refactor-settings * Move Settings, Mode and DEFAULT_SETTINGS to settings.ts
