+++ src/containers/Admin/EditTab/EditTab.tsx
@@ -17,7 +17,6 @@ import { Switch, TextField } from '@entur/form'
 import { Tooltip } from '@entur/tooltip'
 import { ValidationInfoIcon } from '@entur/icons'
 import { Button } from '@entur/button'
-import { Mode } from '../../../settings/useFirebaseSettings'
 import { useSettings } from '../../../settings/SettingsProvider'
 import { isMobileWeb, getTranslation } from '../../../utils/utils'
 import { DEFAULT_DISTANCE, DEFAULT_ZOOM } from '../../../constants'
@@ -33,6 +32,7 @@ import { useRealtimeVehicleData } from '../../../logic/use-realtime-vehicle-data
 import { isNotNullOrUndefined } from '../../../utils/typeguards'
 import { useDebounce } from '../../../hooks/useDebounce'
 import { toggleValueInList } from '../../../utils/array'
+import { Mode } from '../../../settings/settings'
 import { StopPlacePanel } from './StopPlacePanel/StopPlacePanel'
 import { BikePanelSearch } from './BikeSearch/BikePanelSearch'
 import { StopPlaceSearch } from './StopPlaceSearch/StopPlaceSearch'
