+++ src/containers/LandingPage/LandingPage.tsx
@@ -5,8 +5,8 @@ import { Heading1, Heading2, Paragraph, Link } from '@entur/typography'
 import { Contrast } from '@entur/layout'
 import { ForwardIcon } from '@entur/icons'
 import { GridContainer, GridItem } from '@entur/grid'
+import { DEFAULT_SETTINGS } from '../../settings/settings'
 import { createSettings } from '../../services/firebase'
-import { DEFAULT_SETTINGS } from '../../settings/useFirebaseSettings'
 import { SearchPanel } from './SearchPanel/SearchPanel'
 import { TypographyCarousel } from './TypographyCarousel/TypographyCarousel'
 import './LandingPage.scss'
