+++ src/settings/useFirebaseSettings.ts
@@ -1,111 +1,15 @@
 import { useCallback, useEffect, useState } from 'react'
 import { useLocation, useMatch } from 'react-router-dom'
 import { DocumentSnapshot, onSnapshot } from 'firebase/firestore'
-import { Coordinates, TransportMode } from '@entur/sdk'
-import {
-    CustomTile,
-    DashboardTypes,
-    Direction,
-    DrawableRoute,
-    Theme,
-} from '../types'
 import { getSettingsReference } from '../services/firebase'
 import { useUser } from '../UserProvider'
-import { DEFAULT_DISTANCE, DEFAULT_ZOOM } from '../constants'
 import {
     FieldTypes,
     persistMultipleFields as persistMultipleFieldsToFirebase,
     persistSingleField as persistSingleFieldToFirebase,
 } from './FirestoreStorage'
 import type { SettingsSetter } from './SettingsProvider'
-
-type Mode = 'bysykkel' | 'kollektiv' | 'sparkesykkel' | 'delebil'
-
-interface Settings {
-    boardName: string
-    coordinates: Coordinates
-    customImageTiles: CustomTile[]
-    customQrTiles: CustomTile[]
-    dashboard: DashboardTypes
-    description: string
-    direction: Direction
-    distance: number
-    fontScale: number
-    hiddenCustomTileIds: string[]
-    hiddenMobilityOperators: string[]
-    hiddenModes: Mode[]
-    hiddenRealtimeDataLineRefs: string[]
-    hiddenRoutes: { [stopPlaceId: string]: string[] }
-    hiddenStations: string[]
-    hiddenStopModes: { [stopPlaceId: string]: TransportMode[] }
-    hiddenStops: string[]
-    hideRealtimeData: boolean
-    hideSituations: boolean
-    hideTracks: boolean
-    hideWalkInfo: boolean
-    isScheduledForDelete: boolean
-    logo: string
-    logoSize: string
-    newStations: string[]
-    newStops: string[]
-    owners: string[]
-    pageRefreshedAt: number
-    permanentlyVisibleRoutesInMap: DrawableRoute[]
-    scooterDistance: { distance: number; enabled: boolean }
-    showCustomTiles: boolean
-    showIcon: boolean
-    showMap: boolean
-    showPrecipitation: boolean
-    showRoutesInMap: boolean
-    showTemperature: boolean
-    showWeather: boolean
-    showWind: boolean
-    theme: Theme
-    zoom: number
-}
-
-const DEFAULT_SETTINGS: Settings = {
-    boardName: '',
-    coordinates: { latitude: 59.91750312241921, longitude: 10.727442837962354 },
-    customImageTiles: [],
-    customQrTiles: [],
-    dashboard: DashboardTypes.Compact,
-    description: '',
-    direction: Direction.STANDARD,
-    distance: DEFAULT_DISTANCE,
-    fontScale: 1,
-    hiddenCustomTileIds: [],
-    hiddenMobilityOperators: [],
-    hiddenModes: ['sparkesykkel'],
-    hiddenRealtimeDataLineRefs: [],
-    hiddenRoutes: {},
-    hiddenStations: [],
-    hiddenStopModes: {},
-    hiddenStops: [],
-    hideRealtimeData: true,
-    hideSituations: false,
-    hideTracks: false,
-    hideWalkInfo: false,
-    isScheduledForDelete: false,
-    logo: '',
-    logoSize: '32px',
-    newStations: [],
-    newStops: [],
-    owners: [],
-    pageRefreshedAt: 0,
-    permanentlyVisibleRoutesInMap: [],
-    scooterDistance: { distance: DEFAULT_DISTANCE, enabled: false },
-    showCustomTiles: false,
-    showIcon: true,
-    showMap: false,
-    showPrecipitation: true,
-    showRoutesInMap: true,
-    showTemperature: true,
-    showWeather: false,
-    showWind: true,
-    theme: Theme.DEFAULT,
-    zoom: DEFAULT_ZOOM,
-}
+import { DEFAULT_SETTINGS, Settings } from './settings'
 
 export function useFirebaseSettings(): [Settings | null, SettingsSetter] {
     const [settings, setLocalSettings] = useState<Settings | null>(null)
@@ -191,6 +95,3 @@ export function useFirebaseSettings(): [Settings | null, SettingsSetter] {
 
     return [settings, setSettings]
 }
-
-export type { Settings, Mode }
-export { DEFAULT_SETTINGS }
